Move api calls into their own file
